+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added PyPI, DOI and license badges to README.md
 
+### Changed
+
+- Pick a free proxy port when none is specified - [#4](https://github.com/CLOUD-NES/cloudgeometer/pull/4)
+
 ### Fixed
 
 - rasterio reader now properly sets AWS credentials [#2](https://github.com/CLOUD-NES/cloudgeometer/pull/2)